% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_lcbc_logo}
\alias{add_lcbc_logo}
\title{Add LCBC logo to plot}
\usage{
add_lcbc_logo(type = "main", alpha = 0.4, ...)
}
\arguments{
\item{type}{type of logo to get}

\item{alpha}{opacity of logo}

\item{...}{arguments to \code{annotation_custom}}
}
\description{
in a ggplot, adding the logo should be added
early in the plot building
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = disp, colour = cyl)) +
  geom_point() +
  add_lcbc_logo()
}
