% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{lcbc_logo_grob}
\alias{lcbc_logo_grob}
\title{LCBC logo graphics object}
\usage{
lcbc_logo_grob(type = "main", alpha = 0.4)
}
\arguments{
\item{type}{type of logo to get}

\item{alpha}{opacity of logo}
}
\value{
graphics object (grob)
}
\description{
LCBC logo graphics object
}
