% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{lifebrain_cols}
\alias{lifebrain_cols}
\title{Function to extract lifebrain colors as hex codes}
\usage{
lifebrain_cols(...)
}
\arguments{
\item{...}{Character names of lifebrain_colors}
}
\description{
Function to extract lifebrain colors as hex codes
}
