% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_lifebrain}
\alias{scale_color_lifebrain}
\title{Color scale constructor for lifebrain colors}
\usage{
scale_color_lifebrain(...)
}
\arguments{
\item{...}{arguments to \code{\link{scale_colour_lifebrain}}}
}
\description{
Color scale constructor for lifebrain colors
}
