% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_lifebrain}
\alias{theme_lifebrain}
\title{ggplot minimal theme for lifebrain}
\usage{
theme_lifebrain(base_size = base_size_text)
}
\arguments{
\item{base_size}{text size}
}
\description{
ggplot minimal theme for lifebrain
}
