% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_lifebrain_dark_grid}
\alias{theme_lifebrain_dark_grid}
\title{ggplot dark theme for lifebrain}
\usage{
theme_lifebrain_dark_grid(base_size = base_size_text)
}
\arguments{
\item{base_size}{text size}
}
\description{
ggplot dark theme for lifebrain
}
