% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_lifebrain_void}
\alias{theme_lifebrain_void}
\title{ggplot void theme for lifebrain}
\usage{
theme_lifebrain_void(base_size = base_size_text)
}
\arguments{
\item{base_size}{text size}
}
\description{
ggplot void theme for lifebrain
}
